requirements.txt
discord.py
BeatPy